backend/requirements.txt
isort==4.3.21
croniter==0.3.34
cryptography==2.9.2
redis==2.10.6
pymysql==0.9.2
celery==3.1.22
celery-with-redis==3.0
uwsgi==2.0.17.1
gunicorn==19.9.0
channels==2.2.0
python-ldap==3.2.0
paramiko==2.7.1
django==2.0.8
django-celery==3.2.2
djangorestframework-bulk==0.2.1
django-guardian==1.4.9
djangorestframework==3.8.2
django-filter==2.0.0
django-rest-swagger==2.1.1
djangorestframework-jwt==1.11.0
django-cors-headers==2.4.0